Make RendererCancellationThrottle timeout adjustable, and reduce hangs
RendererCancellationThrottle was introduced to allow navigation
cancellations triggered by the same task to work like it did before
even with RenderDocument enabled (see class-level comment for details).

However, we're now suspecting this class is causing increased latency
and reported renderer hangs:
- The throttle has a timeout of 11s is overkill (we can cover 99% of
navs with just 1-2s)
- The navigation is not resumed when the timeout is reached, and
reports the renderer is unresponsive instead.

This CL makes it so that the timeout adjustable and the navigation
is resumed when the timeout is hit. We'll also stop reporting the
timeouts as renderer being unresponsive. If the renderer is truly
hanging, we already have a commit timer as well that will report
the unresponsiveness at commit time (it also has a higher timeout of
30s).

The changes are behind a new flag, so that we can roll it out
safely and compare within the RenderDocument experiment.

Bug: 40615943
Change-Id: I025e89f598e89879b5d75d67398c2985e756faad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7040906
Reviewed-by: Chris Thompson <cthomp@chromium.org>
Reviewed-by: Fergal Daly <fergal@chromium.org>
Auto-Submit: Rakina Zata Amni <rakina@chromium.org>
Commit-Queue: Chris Thompson <cthomp@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530261}