Move the history sync optin flow to the right service on profile swap
During the history sync optin flow, a managed user may select to
create a new managed profile.

In this scenario we resume the history sync flow from the history
sync optin service tied to the newly created profile and stop the
original flow (in the originating profile).

In the passing: we fix the handling of the `management_status_state_`
in ResumeShowHistorySyncOptinScreenFlow: when the updated state
is kManagementDisclaimerComplete, the flow should exit early as we
expect that a follow up state in ResumeShowHistorySyncOptinScreenFlow
will be made where the state will be already marked as complete.

The previous logic resulted in double calling
`ShowHistorySyncOptinScreen` on which we add a sanity CHECK to ensure
a single invocation.

NO_IFTTT=The change does not impact the histogram recording logic, no
changes are necessary.

Bug: 448890655, 452313094
Change-Id: I7a283d17cf0baa36011b4c64340197487f2ab7de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7036754
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Anthi Orfanou <anthie@google.com>
Cr-Commit-Position: refs/heads/main@{#1530264}