Add logging and simplify assertion leading to incognito mode.
Local supervision is distinct from other modes and should not be
considered in the same branch. Moreover, the condition that checks
!IsLocalBrowserFilteringEnabled() requires
IsLocalSearchFilteringEnabled() to be true in order for
IsSupervisedLocally() to be also true; but that won't happen in
kSupervisionEnabledByFamilyLinkPolicy and
kSupervisionEnabledByFamilyLinkUser branches.

Bug: 442625941
Change-Id: I5616125d100491399ce562f94af7667f010eefd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7037761
Commit-Queue: Tomek Jurkiewicz <tju@google.com>
Reviewed-by: Anthi Orfanou <anthie@google.com>
Auto-Submit: Tomek Jurkiewicz <tju@google.com>
Cr-Commit-Position: refs/heads/main@{#1530270}