[Tab] Create a flag to load all tabs at startup
Introduce a flag that will be used to gate creating a WebContents for
every tab at startup on Android. We will not create renderers for
background tabs eagerly, but when enabled this flag will allow us to
reuse more desktop code that expects tabs to always have WebContents.

We speculate that there is a significant performance cost to this
change as in local testing creating a single WebContents even without
a renderer can take 10 ms and use O(100 KiB).

Presently the intent is to 1) develop this feature for evaluation,
2) measure the "real-world" performance cost, 3) profile why the cost
is so high and optimize. Depending on the results we may consider
shipping this change to some configurations.

Bug: 448420873
Change-Id: Ic1a7224a4aa38591fe606e38044230aa2c0552e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7042724
Commit-Queue: Sky Malice <skym@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Auto-Submit: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530283}