[dbsc] Use correct keychain access group on macOS
The keychain access group for the unexportable key service was
hardcoded to the value for Chromium. This is incorrect for branded
builds, such as Google Chrome, which have a different team and bundle
identifier.

This change uses the `GOOGLE_CHROME_BRANDING` build flag to select the
correct keychain access group at build time. This ensures that the
unexportable key service can correctly store and access keys in the
keychain on all macOS builds.

Bug: 384055845, 443931937

Change-Id: I896af5e4080a8a8ab67fc5ddd85d7c4290efc1b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7037318
Reviewed-by: thefrog <thefrog@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530284}