[Predictable Update] Dynamically update chrome://apps on update
This CL ensures that manifest updates are visible on chrome://apps as
soon as they are complete, and doesn't require waiting for a page
refresh to be loaded. This is done by adding a new updateApp() API in
the mojom layer so that the C++ side can communicate an update to the JS
side. Using addApp() also works in case of icon updates only, but if the
name has changed, it doesn't rearrange itself automatically since
addApp() only updates in place.

To ensure that the resource url changes and the "new" resource is
updated to bypass the browser's cache, the icon url is appended with a
timestamp of when the app was applied as a query parameter.

Bug: 451990084
Change-Id: I2398b52b8816946c115e05df50d2d1e920efe80d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7037209
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530300}