Fix flaky AnchorElementSetIsNavigationInDomainBrowserTest.TestPointerDownOnAnchor
The test AnchorElementSetIsNavigationInDomainBrowserTest
TestPointerDownOnAnchor is flaky on Windows; stacks show a timeout
during DoDragDrop, which is an unexpected code path for this test. The
two consecutive link clicks in the test can be incorrectly interpreted
as a drag event if a mouse up event does not occur before the second
link click. To prevent this from happening, set the draggable property
of the anchor element to false.

Example stack:
https://luci-milo.appspot.com/ui/p/chromium/builders/try/win-rel/1201577/test-results?q=ID%3Aninja%3A%2F%2Fchrome%2Ftest%3Abrowser_tests%2FAnchorElementSetIsNavigationInDomainBrowserTest.TestPointerDownOnAnchor+V%3Abuilder%3Dwin-rel+V%3Aos%3DWindows-10-19045+V%3Atest_suite%3Dbrowser_tests

Bug: 333458044
Change-Id: I97c092ff7b7d9b2c94b0fc0ad28a8374bcc276d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7037852
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Liang Zhao <lzhao@microsoft.com>
Commit-Queue: Viktoria Zlatinova <vizlatin@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1530312}