[Card Benefits] Deprecate old submitted form event metrics
We are migrating the card benefit form event enums to a new,
benefit-specific enum class and logging to new histograms. Previous CLs
have added the new form event enums and logging, and we are currently
double-logging metrics to both the old and new histograms.
   Old: Autofill.FormEvents.CreditCard.WithBenefits.*
   New: Autofill.FormEvents.CreditCard.Benefits.*

This CL removes the logging for the old "submitted" form event enums.

Fixed: 417228483
Change-Id: I94e8fa1c1b4dcd405347ccd4f698eeeee189f7a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7027054
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Commit-Queue: Fernando Ramirez <ferny@google.com>
Reviewed-by: Vinny Persky <vinnypersky@google.com>
Cr-Commit-Position: refs/heads/main@{#1530347}