[waapi] Avoid computed undefined progress
When setting the timeline of an animation we compute the current
progress based on the old timeline to that we can align the progress
on the new timeline if required. There are various cases where
the current progress could be undefined.  In this case, we should not
be trying to preserve it.

Bug: 440368332
Change-Id: I62685326ae1a313acbeb027aba811f7c4e8b39e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7042167
Reviewed-by: David Awogbemila <awogbemila@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530378}