[iOS] Update customization menu selection state when submenu cancels.
The underlying issue here is that if a submenu cancels, the NTP
background returns to the initial background. However, the selection
state on the main menu doesn't change (e.g. if the user initially
selected something on the main menu, it will remain visually selected
even as the actual NTP background is reset.

To fix this, the HomeBackgroundCustomizationMediator now alerts its
consumer whenever the current background changes, so the consumer
can update any necessary state.

This can also be used for showing the Cancel/Done buttons in the
submenu view controllers. That refactor will happen in a follow-up CL.

Bug: 450241951
Change-Id: I7e464e0801e873aa44b8be2ad01bcc6882294c6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7031263
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/main@{#1530380}