[Vertical Tabs] Refactor add_child_to_node_ and Initialize
Store the add_child_to_node_ callback on the parent only, and allow it
to take an index argument. Changes Initialize flow so that it only
appends the child's view to the parent after the child is initialized
(i.e. all of the child's children will have been added to the child),
so that there won't be excessive re-layouting of parent views.

Fixed: 450304539
Change-Id: Id36234e890c447f784c5a0de437a974dce879e0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7023207
Reviewed-by: Shibalik Mohapatra <shibalik@chromium.org>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Commit-Queue: Charles Meng <charlesmeng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530399}