[ntp][realbox] Show OGB above Realbox when no dropdown is shown
This CL fixes the issue that OGB popups are shown behind the search box.

This is fixed by changing the z-index of OGB's stacking context. It is
set to 1,000 by default, and it is reduced to -1 when a dropdown menu is
displayed so OGB is put behind the scrim layer.

The change was manually verified:
https://screencast.googleplex.com/cast/NDc0NzYwMzcyMjg5NTM2MHw5MjdjMTg0MS1iOQ

Fixed: b:452322130
Change-Id: Ib06b8847a6acbca3811e5cb7fc776c3208c83abd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7046238
Reviewed-by: Riley Tatum <rtatum@google.com>
Commit-Queue: Sho Mizoe <shomizoe@google.com>
Reviewed-by: Roman Arora <romanarora@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530409}