[Tab] Fix bugs with discard based pending load params
Local testing discovered that requesting `getUrl()` or `getTitle()`
could reset the state of the tab to be based on the WebContents even
when a `mPendingLoadParams` was present. Change the code to special
case the updates for title and URL when `mPendingLoadParams` is set.

This correctly results in the UI now showing the expected URL, title,
and favicon in addition to discarding the thumbnail.

Also if `loadUrl` is invoked we should discard the pending load since
the navigation never got applied.

Bug: 449784092
Change-Id: I6068bc346d1e3667d7c1da5d60d710538a56158c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7042614
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Auto-Submit: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530441}