Combine Android sections of flag_descriptions.h for scripting
This CL continues the work to clean up all the feature flag related
code so that automated scripting can consistently perform tasks. Future
CLs will add scripts and continue clean up.

With this CL we combine many of the one-off sections for Android into
a single cohesive block. We will now have 3 major blocks, and in the
next CLs we will combine them all to have a single section and add a
PRESUBMIT to enforce the same (which will align with the expected
approach of the file, see comment at top of file; and much easy to read
and add new items to).

Note: There are some feature flags that start with "Android" in the
name but are included on each platform. We have not moved these.

Bug: 450940657
Change-Id: Ia0ffe21adcdf7b80937cf2a6e57fbc725321bcd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7046007
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Jack Thiesen <jthiesen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530499}