[Harden Pref][Bug Fix] Fix bypass of encrypted pref validation
This CL fixes a vulnerability in PrefHashFilter::FilterOnLoad where
corrupting a single unencrypted preference hash could prevent the
validation of all encrypted preference hashes.

The fix removes the conditional logic that blocked the scheduling of
the deferred task. This ensures that the unencrypted and encrypted
validation paths are always executed independently. A browser test is
added to verify the logic.


Bug: 449170649
Change-Id: I125df5ef9ca8f99ac9cb950b215dbaf97cbebf66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7037555
Commit-Queue: Zack Han <zackhan@chromium.org>
Reviewed-by: Anunoy Ghosh <anunoy@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530522}