Add README to //media/ffmpeg
We've had some "helpful" tips in //third_party/ffmpeg/README.chromium
for about 200 years now, which have accordingly gone stale, possibly
with new species of mold.

That file has to exist for security reasons in the third party repo, but
I've cleaned it up and added a note to view this file instead.

As far as this file goes, I included some notes that I think are helpful
on how to upstream a patch to FFmpeg. I opted not to include most of the
old tips, as they were so outdated they probably aren't worth keeping.

I also think (open to discussion) that we should move the robosushi
documentation here instead of living in a google doc.

Change-Id: If81d20ad8cd547e538091e3db8b05d00b008d731
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7046102
Commit-Queue: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530527}