Remove UpdateNewTabButtonState function from SidePanelCoordinator
This function was added to support a scenario for Lens that is currently
disabled. We are in the process of refactoring some of the side panel
coordinator logic and so this requires a different way to be
implemented, so removing this function now. We will re-introduce this
functionality in a different manner in the future.

Change-Id: I35f232373a538b3c1e2be12daccbe22a23b5d30d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7046009
Reviewed-by: Steven Luong <stluong@chromium.org>
Commit-Queue: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530531}