[DTC][BrowserControls] Fix DTC transition trigger point
TopControlsStacker uses #getTopControlVisibility to determine if a layer
is part of the height calculation. When the tab strip is transition to
show, it should be considered as part of the "new" browser controls
height.

The previous signal we depend on does not work because the
HIDDEN_BY_HEIGHT_TRANSITION is only set when #onHeightTransition is
called. The height re-calculation of browser controls height happened
before this method is getting called, thus the tab strip transition from
hidden -> shown is dropped.

This is a short term patching fix to bring the feature back to
functionally. Fixing the UI presentation (the first frame when tab strip
is hiding) is not part of this CL.

Demo (Canary as before the fix)
https://screencast.googleplex.com/cast/NTQxMTc5NTA2NDg0ODM4NHxiYzI0ZjlkNC0zNQ

Bug: 448174907
Change-Id: I0efbe1695539590f5d0bd4d17d711fb6e745fa89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7045949
Commit-Queue: Wenyu Fu <wenyufu@chromium.org>
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530534}