net/disk_cache: Add option to disable synchronous writes in SQL cache
This change introduces a new feature parameter,
kSqlDiskCacheSynchronousOff, to control the SQLite synchronous
pragma when the SQL disk cache is operating in WAL mode.

When this flag is enabled, the pragma is set to OFF. In this mode,
SQLite hands data to the OS for writing but does not wait for the
write to be durable. This can significantly improve performance, but
carries a risk of database corruption in the event of an OS crash or
power failure. However, with data checksums introduced in
crrev.com/c/7023771, it is unlikely that a resource for a different
URL would be served even if corruption occurs.

The default behavior remains PRAGMA synchronous = NORMAL, which ensures
that the database will not be corrupted after a power failure.

This is added as part of the kDiskCacheBackendExperiment to evaluate
the performance vs. reliability trade-off.

Performance was measured by running the
`DiskCachePerfTest.SqlCacheBackendPerformance` test 10 times and taking
the median on an M4 Pro MacBook Pro (16-inch, Nov 2024).

|                              | ToT r1529907 | sync = OFF   |
| -----------------------------|--------------|--------------|
| cache_entries_write_time     | 2,083.156 ms | 1,483.769 ms |
| cache_headers_read_time_cold |   626.414 ms |   518.622 ms |
| cache_headers_read_time_warm |   285.133 ms |   195.459 ms |
| cache_entries_read_time_cold | 1,251.832 ms | 1,064.859 ms |
| cache_entries_read_time_warm |   394.465 ms |   316.271 ms |

Bug: 436446789, 443171275
Change-Id: I3dd2a86fdea7b1cd42f46d180426010a069bc44d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6869309
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Mingyu Lei <leimy@chromium.org>
Reviewed-by: Olivier Li <olivierli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530639}