FindInPage: Update FindBuffer::BufferNodeMapping
This CL is a preparation to fix find-in-page for <textarea>.
We're going to support a buffer string consisted of multiple IFCs.

* Add `const Member<const OffsetMapping>` data member to
  FindBuffer::BufferNodeMapping
  - FindBuffer::AddTextToBuffer() appends the current OffsetMapping
    instance.
  - FindBuffer::PositionAt*OfCharacterAtIndex() refer to it.

* Make FindBuffer::BufferNodeMapping public for
  WTF_ALLOW_CLEAR_UNUSED_SLOTS_WITH_MEM_FUNCTIONS

* Make FindBuffer::buffer_node_mappings_ a HeapVector

The size of FindBuffer::buffer_node_mappings_ is usually small because
the size depends on the number of collapsed character sequences, so we
expect memory usage won't increase drastically by this change.

Bug: 438220615
Change-Id: I4c5627918e233759ae5e32aa5c0023732aa2e4f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7043713
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Auto-Submit: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530644}