[iOS] Tentative fix for toolbar crash
There is a crash where `self.locationBarView` is nil.
It is very unclear how it is possible considering that it was set
right above. But maybe there is an interaction hard to detect.

In anycase, as the rest of the method is using the local variable
and not the property, change the constraints to do the same.

Fixed: 452035899
Change-Id: I4613132054556714152ee51731a6297179e02d96
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7045739
Reviewed-by: Christian Xu <christianxu@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530703}