Update preferences sync test to account for always-syncing prefs
The new always-syncing prefs are not bound to the preferences sync
toggle and thus are not removed when the type is disabled. This calls
for an update to the preferences sync tests which expect all the account
preferences to get removed when prefs sync is turned off.

Fixed: 452499958
Change-Id: I5630192312a6262803c9455aac8149505d68d086
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7048148
Reviewed-by: Marc Treib <treib@chromium.org>
Auto-Submit: Ankush Singh <ankushkush@google.com>
Commit-Queue: Ankush Singh <ankushkush@google.com>
Cr-Commit-Position: refs/heads/main@{#1530710}