[ios] Mitigate blank snapshots taken before the navigation has committed
This CL mitigates the issue of blank snapshots taken before the
navigation has committed. Prior to this change, the generateSnapshot
method returned early when a last committed url is nil and it makes a
snapshot blank.

With this change, the implementation in Swift allows
lastCommittedUrl.nsurl to be nil and moves forward to take a snapshot.
Note that it's not an ideal solution because the
generateWKWebViewSnapshot method can be used for NTPs when
lastCommittedUrl.nsurl is nil. It potentially generates a wrong snapshot
image. The potential issue has been existed in the Objective-C
implementation for the long time.

Fixed: 451615903
Bug: 452299163
Change-Id: Id8e33c82e63a2511bc21222b02f3807b3abb36f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7045417
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Asami Doi <asamidoi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530718}